+++ contracts/switchboard/default-switchboards/FastSwitchboard.sol
@@ -25,7 +25,7 @@ contract FastSwitchboard is SwitchboardBase {
     // Event emitted when a new socket is set
     event SocketSet(address newSocket);
     // Event emitted when a root is attested
-    event ProposalAttested(bytes32 packetId, uint64 proposalId, bytes32 root, address attester);
+    event ProposalAttested(bytes32 packetId, uint256 proposalId, bytes32 root, address attester);
 
     // Error emitted when a watcher is found
     error WatcherFound();
@@ -69,7 +69,7 @@ contract FastSwitchboard is SwitchboardBase {
      * @param proposalId_ Proposal ID
      * @param signature_ Signature of the packet
      */
-    function attest(bytes32 packetId_, uint64 proposalId_, bytes calldata signature_) external {
+    function attest(bytes32 packetId_, uint256 proposalId_, bytes calldata signature_) external {
         uint32 srcChainSlug = uint32(uint256(packetId_) >> 224);
         bytes32 root = socket__.packetIdRoots(packetId_, proposalId_);
         if (root == bytes32(0)) revert InvalidRoot();
@@ -95,12 +95,13 @@ contract FastSwitchboard is SwitchboardBase {
     /**
      * @notice verifies if the packet satisfies needed checks before execution
      * @param packetId_ packetId
+     * @param proposalId_ proposalId
      * @param proposeTime_ time at which packet was proposed
      */
     function allowPacket(
         bytes32 root_,
         bytes32 packetId_,
-        uint64 proposalId_,
+        uint256 proposalId_,
         uint32 srcChainSlug_,
         uint256 proposeTime_
     ) external view override returns (bool) {
